requirements.txt
requests==2.31.0
boto3==1.34.34
python-dotenv==1.0.0
psycopg2-binary==2.9.9

pandas==2.1.4
openpyxl==3.1.2